Laporan 3 (WEEK 9) Final Tugas Besar (Mobile App)/backend/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
requests
firebase-admin
